more fixing

+++ src/lib/coaler/multialign/scorer/AlignmentScorer.cpp
@@ -6,7 +6,7 @@
 
 namespace coaler::multialign {
     double AlignmentScorer::calcTanimotoShapeSimilarity(const RDKit::ROMol &molA, const RDKit::ROMol &molB,
-                                                           unsigned int posIdA, unsigned int posIdB) {
-        return 1 - RDKit::MolShapes::tanimotoDistance(molA, molB, posIdA, posIdB);
+                                                        unsigned int posIdA, unsigned int posIdB) {
+        return 1 - RDKit::MolShapes::tanimotoDistance(molA, molB, static_cast<int>(posIdA), static_cast<int>(posIdB));
     }
 }  // namespace coaler::multialign
